requirements.txt
Flask
numpy
opencv-python
python-dotenv
torch
torchvision
Pillow
tensorflow
google-generativeai
markdown
beautifulsoup4
